# Optimize retrieval of simple name from Assembly.FullName, fix up faulty methods