--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/documents/DocumentsTrace.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/documents/DocumentsTrace.cs
@@ -117,7 +117,7 @@ namespace MS.Internal.Documents
         public DocumentsTrace(string switchName)
         {
 #if DEBUG
-            string name = SafeSecurityHelper.GetAssemblyPartialName( Assembly.GetCallingAssembly() );
+            string name = ReflectionUtils.GetAssemblyPartialName(Assembly.GetCallingAssembly());
             _switch = new BooleanSwitch(switchName, $"[{name}]");
 #endif
         }