--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Markup/Primitives/MarkupWriter.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Markup/Primitives/MarkupWriter.cs
@@ -1,17 +1,17 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
 
 //
 //  Contents:  XAML writer
 //
 
+using System.Xml.Serialization;
 using System.ComponentModel;
-using System.Reflection;
 using System.Collections;
+using System.Reflection;
+using MS.Internal;
 using System.Text;
 using System.Xml;
-using System.Xml.Serialization;
 
 namespace System.Windows.Markup.Primitives
 {