--- a/src/Microsoft.DotNet.Wpf/src/Shared/MS/Internal/ReflectionUtils.cs
+++ b/src/Microsoft.DotNet.Wpf/src/Shared/MS/Internal/ReflectionUtils.cs
@@ -16,7 +16,6 @@ namespace MS.Internal
     /// </summary>
     internal static class ReflectionUtils
     {
-
 #if !NETFX
         /// <summary>
         ///  Retrieves the full assembly name by combining the <paramref name="partialName"/> passed in