--- a/src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/SplashScreen.cs
+++ b/src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/SplashScreen.cs
@@ -53,9 +53,8 @@ namespace System.Windows
             }
 
             _resourceName = resourceName.ToLowerInvariant();
-            _hInstance = (HINSTANCE)Marshal.GetHINSTANCE(resourceAssembly.ManifestModule);
-            AssemblyName name = new(resourceAssembly.FullName);
-            _resourceManager = new ResourceManager($"{name.Name}.g", resourceAssembly);
+            _hInstance = Marshal.GetHINSTANCE(resourceAssembly.ManifestModule);
+            _resourceManager = new ResourceManager($"{ReflectionUtils.GetAssemblyPartialName(resourceAssembly)}.g", resourceAssembly);
         }
 
         public void Show(bool autoClose)