… src/Microsoft.DotNet.Wpf/src/WindowsBase/WindowsBase.csproj