## Description

Optimizes retrieval of simple name from `Assembly.FullName` and removes all the weird ways used throughout to get it, this optimizes both startup and runtime performance.
- In `Baml2006Reader` and `XamlSchemaContext` the original method did not properly parse assembly names with properly escaped commas, a weird edge-case but still a valid one. This was there due to CAS anyways.
- Since we're in all cases as far as I know working with `RuntimeAssembly`, using `assembly.GetName().Name` is always the slowest, followed by `new AssemblyName(assembly.FullName).Name`.
- The original intention was to merely unify this behaviour, however given how slow it is and since there was already custom parsing done in Xaml/Baml, I've decided to roll our own.
- The two methods were moved from `SafeSecurityHelper` to its own `ReflectionUtils` static class.
- Since the input is trusted (`FullName` is already after `AssemblyName` parser sanization, and with `RuntimeAssembly` (even dynamic assemblies from `AssemblyBuilder`), we can just look for a separator, there should be no concerns with custom parsing.
- PBT on `net472` fallbacks to `new AssemblyName(assembly.FullName).Name`.

### Benchmarks for RuntimeAssembly

| Method       | assembly             | Mean [ns] | Error [ns] | StdDev | Gen0   | Code Size | Allocated |
|------------- |--------------------- |----------:|-----------:|--------:|-------:|----------:|----------:|
| GetName_Name | DataF(...)=null [75] | 762.18 ns |   4.407 ns |    4.123 ns | 0.0439 |       1,768 B |         736 B |
| AssemblyName | DataF(...)=null [75] | 382.63 ns |   4.307 ns |    4.028 ns | 0.0291 |         3,140 B |         488 B |
| PR_EDIT      | DataF(...)=null [75] |  10.25 ns |   0.045 ns |    0.038 ns |      - |       1,629 B |             - |
| GetName_Name | Syste(...)7798e [89] | 516.84 ns |   4.869 ns |    4.555 ns | 0.0362 |       1,761 B |         616 B |
| AssemblyName | Syste(...)7798e [89] | 387.02 ns |   3.930 ns |    3.676 ns | 0.0310 |         3,115 B |         520 B |
| PR_EDIT      | Syste(...)7798e [89] |  10.75 ns |   0.051 ns |    0.045 ns |      - |       1,629 B |             - |
| GetName_Name | Bench(...)cefc4 [84] | 739.83 ns |   7.599 ns |    6.736 ns | 0.0486 |       1,761 B |         816 B |
| AssemblyName | Bench(...)cefc4 [84] | 375.59 ns |   4.439 ns |    4.153 ns | 0.0300 |         3,140 B |         504 B |
| PR_EDIT      | Bench(...)cefc4 [84] |  10.76 ns |   0.078 ns |    0.073 ns |      - |       1,629 B |             - |

### Benchmark code
<details><summary>Benchmark code</summary>

```csharp
public static IEnumerable<Assembly> AssembliesToTest
{
    get
    {
        // DataFormatsBenchmark, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null
        yield return AppDomain.CurrentDomain.GetAssemblies()[1];
        // System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e
        yield return AppDomain.CurrentDomain.GetAssemblies()[0];
        // BenchmarkDotNet, Version=0.13.12.0, Culture=neutral, PublicKeyToken=aa0ca2f9092cefc4
        yield return AppDomain.CurrentDomain.GetAssemblies()[2];
    }
}

[Benchmark]
[ArgumentsSource(nameof(AssembliesToTest))]
public string AssemblyName(Assembly assembly)
{
    AssemblyName name = new AssemblyName(assembly.FullName);
    string partialName = name.Name;
    return (partialName != null) ? partialName : string.Empty;
}

[Benchmark]
[ArgumentsSource(nameof(AssembliesToTest))]
public string GetName_Name(Assembly assembly)
{
    return assembly.GetName().Name ?? string.Empty;
}

[Benchmark]
[ArgumentsSource(nameof(AssembliesToTest))]
public ReadOnlySpan<char> PR_EDIT(Assembly assembly)
{
    ReadOnlySpan<char> fullName = assembly.FullName;
    if (fullName.IsEmpty)
        return ReadOnlySpan<char>.Empty;

    ReadOnlySpan<char> nameSlice = fullName;
    // Skip any escaped commas in the name if present
    int escapedComma = fullName.LastIndexOf("\\,", StringComparison.Ordinal);
    if (escapedComma != -1)
        nameSlice = nameSlice.Slice(escapedComma + 2);

    // Find the real ending of the name section
    int commaIndex = nameSlice.IndexOf(',');
    if (commaIndex != -1)
        fullName = fullName.Slice(0, fullName.Length - nameSlice.Length + commaIndex);

    // Check if we need to unescape, this is very rare case so we can just do it the dirty way
    if (escapedComma != -1 || fullName.Contains('\\'))
        UnescapeDirty(ref fullName);

    // Since having "," or "=" in the assembly name is very rare, we don't want to inline
    [MethodImpl(MethodImplOptions.NoInlining)]
    static void UnescapeDirty(ref ReadOnlySpan<char> dirtyName)
    {
        StringBuilder sb = new(dirtyName.Length);
        sb.Append(dirtyName);
        sb.Replace("\\", null);
        dirtyName = sb.ToString();
    }

    return fullName;
}
```

</details>

## Customer Impact

Improved startup/runtime performance, decreased allocations.

## Regression

No.

## Testing

Local build, sample app run, assert tests.

```csharp
// Please put "allows ref struct" here, those ToString() hurt
// System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e
// Expected "System.Private.CoreLib"
ReadOnlySpan<char> x1 = GetAssemblyPartialName(AppDomain.CurrentDomain.GetAssemblies()[0]);
Assert.AreEqual(x1.ToString(), AppDomain.CurrentDomain.GetAssemblies()[0].GetName().Name);

// Some additional tests to test consistency of the parser
AssemblyBuilder asm1 = AssemblyBuilder.DefineDynamicAssembly(new AssemblyName("1"), AssemblyBuilderAccess.Run);
AssemblyBuilder asm2 = AssemblyBuilder.DefineDynamicAssembly(new AssemblyName("\\,1\\="), AssemblyBuilderAccess.Run);
AssemblyBuilder asm3 = AssemblyBuilder.DefineDynamicAssembly(new AssemblyName("\\,thisis\\,alsovalid.name"), AssemblyBuilderAccess.Run);
AssemblyBuilder asm4 = AssemblyBuilder.DefineDynamicAssembly(new AssemblyName("DataFormat\\,\\, Benchmark\\, Version\\=, Culture=neutral, PublicKeyToken=null"), AssemblyBuilderAccess.Run);
AssemblyBuilder asm5 = AssemblyBuilder.DefineDynamicAssembly(new AssemblyName("DataFormat\\,\\, Benchmark\\, Version\\=, Culture=neutral, PublicKeyToken=null"), AssemblyBuilderAccess.Run);
AssemblyBuilder asm6 = AssemblyBuilder.DefineDynamicAssembly(new AssemblyName("\\,\\=+thisis\\,alsovalid.name"), AssemblyBuilderAccess.Run);
x1 = GetAssemblyPartialName(asm1);
Assert.AreEqual(x1.ToString(), asm1.GetName().Name);
x1 = GetAssemblyPartialName(asm2);
Assert.AreEqual(x1.ToString(), asm2.GetName().Name);
x1 = GetAssemblyPartialName(asm3);
Assert.AreEqual(x1.ToString(), asm3.GetName().Name);
x1 = GetAssemblyPartialName(asm4);
Assert.AreEqual(x1.ToString(), asm4.GetName().Name);
x1 = GetAssemblyPartialName(asm5);
Assert.AreEqual(x1.ToString(), asm5.GetName().Name);
x1 = GetAssemblyPartialName(asm6);
Assert.AreEqual(x1.ToString(), asm6.GetName().Name);
```

## Risk

In my mind it is low, all places I've tested are passing `RuntimeAssembly`, so there should be no surprises.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9739)