@h3xds1nz 2025-01-15
@himgoyalmicro Awesome, now just the other one and I can continue the journey <3

Merge conflicts should be solved in the last commit, was kind of few wrong auto-merges of usings. 

Regarding tests, yeah, I'll get a PR in somewhere this week.