@ThomasGoulet73 2025-01-15
Sorry for being a bit late but wouldn't it be better to use the new [AssemblyNameInfo](https://learn.microsoft.com/en-us/dotnet/api/system.reflection.metadata.assemblynameinfo) instead ? It should be faster than the existing code using AssemblyName and it might be slower than the custom parser that you implemented but it would be better than using our own parser that might contain bugs and that the WPF team would need to maintain.