@h3xds1nz 2025-01-15
@ThomasGoulet73 It uses the same parser impl basically except it doesn't throw because it has `TryParse`. I've considered it originally.

As I've mentioned, I wouldn't use custom impl if it wasn't already parsed by the runtime parser. It is not a science and it is well documented what you need to escape. Plus WPF was already doing custom slicing and it was doing it wrong.