@ThomasGoulet73 2025-01-17
I mean I wouldn't call it a "science" but I wouldn't call it straightforward either. The thing is your implementation does not do exactly the same thing as GetName/AssemblyName. Assembly names can contain chars such as '\t', '\n', '\r' or '\\\\'. You can see in your own test that if you use `new AssemblyName("Test\n\Test")` the result differs between your parser and GetName/AssemblyName/AssemblyNameInfo.

It most probably never happen in practice but I just wanted to make it known that it wasn't a 1-to-1 replacement.

Also FYI I wrote a quick fuzzer to test it out:
<details>
  <summary>Code</summary>
  
```
char[] supportedChars = ['a', 'b', 'c', '\\', ',', '=', '\'', '"', 't', 'r', 'n'];

for (int x = 0; x < 10_000; x++)
{
    int len = Random.Shared.Next(1, 30);

    char[] chars = new char[len];

    for (int y = 0; y < chars.Length; y++)
    {
        int i = Random.Shared.Next(0, supportedChars.Length);

        chars[y] = supportedChars[i];
    }

    var str = new string(chars);

    if (AssemblyNameInfo.TryParse(str, out AssemblyNameInfo ani))
    {
        var name = GetAssemblyPartialName(ani.FullName);

        if (name.ToString() != ani.Name)
        {
            File.AppendAllLines("Test1.txt", [$"{ani.FullName}: {name.ToString()} != {ani.Name}"]);
        }
    }
}
```
  
</details>

> It uses the same parser impl basically except it doesn't throw because it has `TryParse`.

That's true. It's not much faster but it allocates 1 less instance (AssemblyName itself), can be used with span and won't instantiate a CultureInfo if Culture != neutral.